#2 : detach access to user_oauth_info table logic
